feat(cli): add choosing repository on install
